add hash attribute to assure unicity of images in database; add unit tests for upload method and GeorefImage mapping;

+++ src/main/java/com/georeso/georef_drawing_service/georef/service/GeorefImageService.java
@@ -4,7 +4,6 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.time.LocalDateTime;
-import java.util.UUID;
 
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -28,30 +27,32 @@ public class GeorefImageService {
     private final StorageConfig storageConfig;
 
     public GeorefImageDto uploadImage(MultipartFile file) {
-    try {
-        String hash = FileUtils.calculateSHA256(file);
+        try {
+            String hash = FileUtils.calculateSHA256(file);
 
-        // Vérifier si une image avec le même hash existe déjà
-        if (repository.existsByHash(hash)) {
-            return GeorefMapper.toDto(repository.findByHash(hash), null, null);
+            // Vérifier si une image avec le même hash existe déjà
+            if (repository.existsByHash(hash)) {
+                return GeorefMapper.toDto(repository.findByHash(hash), null, null);
+            }
+
+            String filename = hash + "_" + file.getOriginalFilename();
+            Path targetPath = storageConfig.getOriginalDir().resolve(filename);
+            if (!Files.exists(targetPath)) {
+                Files.copy(file.getInputStream(), targetPath);
+            }
+
+            GeorefImage image = new GeorefImage();
+            image.setFilepathOriginal(targetPath.toString());
+            image.setHash(hash);
+            image.setUploadingDate(LocalDateTime.now());
+            image.setStatus(GeorefStatus.UPLOADED);
+
+            GeorefImage savedImage = repository.save(image);
+            return GeorefMapper.toDto(savedImage, null, null);
+
+        } catch (IOException e) {
+            throw new ImageUploadException("Impossible d'importer l'image", e);
         }
-
-        String filename = UUID.randomUUID() + "_" + file.getOriginalFilename();
-        Path targetPath = storageConfig.getOriginalDir().resolve(filename);
-        Files.copy(file.getInputStream(), targetPath);
-
-        GeorefImage image = new GeorefImage();
-        image.setFilepathOriginal(targetPath.toString());
-        image.setHash(hash);
-        image.setUploadingDate(LocalDateTime.now());
-        image.setStatus(GeorefStatus.UPLOADED);
-
-        GeorefImage saved = repository.save(image);
-        return GeorefMapper.toDto(saved, null, null);
-
-    } catch (IOException e) {
-        throw new ImageUploadException("Impossible d'importer l'image", e);
     }
-}
 
 }
